refactor: fold the C-API and wasm bindings into the main crate as opt-in features
The `graphitesql-capi` and `graphitesql-wasm` sibling crates are merged into
the main crate as feature-gated FFI modules, so there is one crate (and one
version) covering the Rust library, the `libsqlite3`-compatible C ABI, and
the WebAssembly bindings.

- `#![forbid(unsafe_code)]` → `#![deny(unsafe_code)]` (and `[lints.rust]`
  likewise). `forbid` can't be locally overridden, which is why the shims had
  to be separate crates; `deny` lets the two FFI modules carry a scoped
  `#[allow(unsafe_code)]`. The engine remains 100% unsafe-free, and the default
  build (neither feature) contains no `unsafe` and no FFI.
- `capi` feature → `src/capi.rs` (the 89-symbol `sqlite3_*` C ABI; pulls `std`).
  `wasm` feature → `src/wasm.rs` (wasm-bindgen + OPFS VFS; adds the optional
  `wasm-bindgen`/`js-sys`/`web-sys` deps). Both off by default.
- The `[lib]` crate-type stays `rlib`, so `cargo build --no-default-features`
  is still a bare no_std library (a no_std cdylib/staticlib would need a global
  allocator + panic handler). The C shared/static lib and the wasm module are
  produced on demand via `cargo rustc --lib --features <f> --crate-type
  cdylib|staticlib` (the purecrypto FFI pattern).
- Ported assets: `include/sqlite3.h`, `tests/capi/{ctest.c,run.sh}` (the C
  consumer test, rewired to link `-lgraphitesql`), `examples/wasm/`,
  `tests/wasm_node_smoke.mjs`. CI's wasm/capi jobs now build from the main crate.
- Portability fix surfaced by wasm32 clippy: `jsonb_header_len`/the JSONB
  encoder compared `usize <= 0xFFFF_FFFF`, always-true on 32-bit; cast to `u64`.

Verified: default gate clean (fmt/clippy `--all-features`/no_std rlib/doc), the
C consumer test passes against the on-demand cdylib (all `sqlite3_*`), and the
wasm cdylib builds for `wasm32-unknown-unknown`.

+++ src/exec/json.rs
@@ -383,7 +383,7 @@ fn jsonb_header_len(n: usize) -> usize {
         2
     } else if n <= 0xFFFF {
         3
-    } else if n <= 0xFFFF_FFFF {
+    } else if n as u64 <= 0xFFFF_FFFF {
         5
     } else {
         9
@@ -471,7 +471,7 @@ fn push_jsonb(out: &mut Vec<u8>, ty: u8, payload: &[u8]) {
     } else if n <= 0xFFFF {
         out.push((13 << 4) | ty);
         out.extend_from_slice(&(n as u16).to_be_bytes());
-    } else if n <= 0xFFFF_FFFF {
+    } else if n as u64 <= 0xFFFF_FFFF {
         out.push((14 << 4) | ty);
         out.extend_from_slice(&(n as u32).to_be_bytes());
     } else {
